rename EachAssignments
